fix: imports on ec2.py, stats.py, instances.py, vpcs.py and security_groups.py, also add import in main.py and setup tools
