+++ aws_inventory/main.py
@@ -1,4 +1,5 @@
 import argparse
+from tqdm import tqdm
 from aws_inventory.regional.ec2 import collect_ec2
 from aws_inventory.utils.html_report import render_html, save_output
 from aws_inventory.utils.boto_helpers import create_session, get_all_regions
@@ -24,12 +25,14 @@ def main():
     session = create_session(args.profile)
     regions = parse_regions(args.regions, session)
 
+    print(f"\nStarting AWS inventory collection for {len(regions)} region(s)...\n")
+
     # Collect inventories grouped by service type
     inventories_by_service = {}
 
-    # Collect EC2 data for all regions
+    # Collect EC2 data for all regions with progress bar
     ec2_regions_data = {}
-    for region in regions:
+    for region in tqdm(regions, desc="Collecting EC2 data", unit="region"):
         ec2_data = collect_ec2(args.profile, region)
         ec2_regions_data[region] = ec2_data
 
@@ -40,18 +43,14 @@ def main():
             "regions": ec2_regions_data
         }
 
-    # Example for future global services:
-    # iam_data = collect_iam(args.profile)
-    # inventories_by_service["IAM"] = {
-    #     "type": "iam",
-    #     "global": True,
-    #     "data": iam_data
-    # }
+    print("\nGenerating HTML report...")
 
     # Render HTML from structured data
     html_content = render_html(inventories_by_service, args.profile)
     filename = "inventory_report.html"
     save_output(html_content, filename)
+
+    print("Inventory collection complete!\n")
 
 
 if __name__ == "__main__":
