+++ requirements.txt
@@ -1,4 +1,4 @@
 boto3>=1.30.0
 jinja2>=3.1.2
+tqdm>=4.66.0
 python-dateutil>=2.9.0
-rich>=13.5.2
